backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
pymysql==1.1.0  # Alternative MySQL driver
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
selenium==4.15.2
webdriver-manager==4.0.1

# Development Tools
ruff==0.1.6
pre-commit==3.5.0

# Environment Variables
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Data Validation
validators==0.22.0
